requirements.txt
numpy==1.19.2
opencv-python==4.4.0.44
pandas==1.1.2
Pillow==7.2.0
scikit-learn==0.23.2
tensorflow==2.3.1
